refactor(search): move config to its own file
