+++ x-pack/solutions/search/plugins/enterprise_search/server/config.ts
@@ -1,0 +1,47 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { schema, TypeOf } from '@kbn/config-schema';
+import { PluginConfigDescriptor } from '@kbn/core/server';
+
+export const configSchema = schema.object({
+  accessCheckTimeout: schema.number({ defaultValue: 5000 }),
+  accessCheckTimeoutWarning: schema.number({ defaultValue: 300 }),
+  customHeaders: schema.maybe(schema.object({}, { unknowns: 'allow' })),
+  enabled: schema.boolean({ defaultValue: true }),
+  hasConnectors: schema.boolean({ defaultValue: true }),
+  hasDefaultIngestPipeline: schema.boolean({ defaultValue: true }),
+  hasDocumentLevelSecurityEnabled: schema.boolean({ defaultValue: true }),
+  hasIncrementalSyncEnabled: schema.boolean({ defaultValue: true }),
+  hasNativeConnectors: schema.boolean({ defaultValue: true }),
+  hasWebCrawler: schema.boolean({ defaultValue: false }),
+  host: schema.maybe(schema.string()),
+  isCloud: schema.boolean({ defaultValue: false }),
+  ssl: schema.object({
+    certificateAuthorities: schema.maybe(
+      schema.oneOf([schema.arrayOf(schema.string(), { minSize: 1 }), schema.string()])
+    ),
+    verificationMode: schema.oneOf(
+      [schema.literal('none'), schema.literal('certificate'), schema.literal('full')],
+      { defaultValue: 'full' }
+    ),
+  }),
+  ui: schema.object({
+    enabled: schema.boolean({ defaultValue: true }),
+  }),
+});
+
+export type ConfigType = TypeOf<typeof configSchema>;
+
+export const config: PluginConfigDescriptor<ConfigType> = {
+  deprecations: ({ unused }) => [unused('canDeployEntSearch', { level: 'warning' })],
+  exposeToBrowser: {
+    host: true,
+    ui: true,
+  },
+  schema: configSchema,
+};
